+++ Runtime/Plugins/iOS/ACPaymentLinks.xcframework/ios-arm64/ACPaymentLinks.framework/Headers/ACPaymentLinks.h
@@ -14,3 +14,4 @@ FOUNDATION_EXPORT double ACPaymentLinksVersionNumber;
 FOUNDATION_EXPORT const unsigned char ACPaymentLinksVersionString[];
 
 #import <ACPaymentLinks/ACBridgeAPI_Bridge.h>
+#import <SafariServices/SafariServices.h>